+++ src/environments/environment.prod.ts
@@ -2,7 +2,7 @@ export const environment = {
   production: true,
   endpoint: '/api/v1',
   discordInvite: 'https://discord.gg/uDTgxyg',
-  version: 'v0.1.1b',
+  version: 'v0.1.2b',
   url: 'https://dbots.co',
   guildId: '236608364333891585'
 };