+++ src/environments/environment.ts
@@ -6,7 +6,7 @@ export const environment = {
   production: false,
   endpoint: 'http://localhost:3000/api/v1',
   discordInvite: 'https://discord.gg/uDTgxyg',
-  version: 'v0.1.1b',
+  version: 'v0.1.2b',
   url: 'http://localhost:3000',
   guildId: '531196495584821314'
 };